perf(recon): write CDEF interior output as u16 (#959)
Feature: INFRA-DECODE-PARALLEL-STAGES

+++ crates/splot-decode/src/filters/cdef.rs
@@ -603,10 +603,10 @@ fn compute_cdef_filter_plane<T: ReconSample>(
             dir: ctx.dir,
             coeff_shift: ctx.coeff_shift,
         };
-        let mut out = [0i32; 64];
+        let mut out = [0u16; 64];
         cdef_filter_block_interior(&pad, w, h, &filter, &mut out);
         for (dst, &filtered) in filtered_block.iter_mut().zip(&out).take(w * h) {
-            *dst = storage_sample::<T>(filtered, ctx.max_sample)?;
+            *dst = storage_sample::<T>(i32::from(filtered), ctx.max_sample)?;
         }
     } else {
         let offsets = CdefTapOffsets::for_direction(ctx.dir);
